in progress

+++ src/modules/docs/docs.js
@@ -3,67 +3,45 @@ import { getParam } from '../../js/history';
 import * as API from '../../js/api';
 import Auth from '../../js/Auth';
 
-const $form = $('#form-docs');
 const types = [
   'TECHNICAL_DOCUMENT',
   'LEGAL_DOCUMENT'
 ];
 
-if ($form.length) {
-  const orderId = getParam('order') || getParam('reference');
+export default function processOrder(order) {
+  const $form = $('#form-docs');
+  const isEdit = $form.data('edit');
 
-  Auth.getProfile()
-    .then(loadOrder)
-    .catch(() => {
-      Auth.showLoginPopup().then(
-        loadOrder,
-        () => (window.location = '.')
-      );
-    });
+  const id = getParam('id');
+  const operation = getParam('operation');
+  const reference = getParam('reference');
+  if (id && operation && reference) API.confirmPayment(id, operation, reference, true, Auth.token);
 
-  function loadOrder() {
-    const orderId = getParam('order') || getParam('reference');
-    API.getDraft(orderId, Auth.token).then(
-      processOrder,
-      () => {
-        $('.form__spinner').hide();
-        $('.form__error').show();
-      }
-    );
-  }
+  $('.form__spinner').hide();
+  $('.form__form').show();
+  $('.docs__error').html('');
 
-  function processOrder(order) {
-    const id = getParam('id');
-    const operation = getParam('operation');
-    const reference = getParam('reference');
-    if (id && operation && reference) API.confirmPayment(id, operation, reference, true, Auth.token);
+  $form.on('submit', (e) => {
+    e.preventDefault();
+    // const data = {
+    //   id:                 orderId,
+    //   legalPersonOwner:   $('#form-legalPersonOwner').prop('checked'),
+    //   minorOwner:         $('#form-minorOwner').prop('checked'),
+    //   onerousTransaction: $('#form-onerousTransaction').prop('checked')
+    // };
+    API.createOrder(order.id, Auth.token)
+      .then(() => (window.location.href = $form.attr('action')))
+      .catch(err => $('.docs__error').html(err.responseJSON.error));
+  });
 
-    $('.form__spinner').hide();
-    $('.form__form').show();
-    $('.docs__error').html('');
-
-    $form.on('submit', (e) => {
-      e.preventDefault();
-      // const data = {
-      //   id:                 orderId,
-      //   legalPersonOwner:   $('#form-legalPersonOwner').prop('checked'),
-      //   minorOwner:         $('#form-minorOwner').prop('checked'),
-      //   onerousTransaction: $('#form-onerousTransaction').prop('checked')
-      // };
-      API.createOrder(orderId, Auth.token)
-        .then(() => (window.location.href = $form.attr('action')))
-        .catch(err => $('.docs__error').html(err.responseJSON.error));
-    });
-
-    if (Array.isArray(order.attachedFileList)) {
-      types.forEach((type) => {
-        const $el = $(`#${type}`);
-        order.attachedFileList.forEach((file) => {
-          if (file.fileType !== type) return;
-          $el.append(createFile(file));
-        });
+  if (Array.isArray(order.attachedFileList)) {
+    types.forEach((type) => {
+      const $el = $(`#${type}_LIST`);
+      order.attachedFileList.forEach((file) => {
+        if (file.fileType !== type) return;
+        $el.append(createFile(file));
       });
-    }
+    });
   }
 
   require.ensure([], () => {
@@ -75,7 +53,7 @@ if ($form.length) {
       const $bar = $el.find('.docs__progressbar');
 
       $el.find('.docs__input').fileupload({
-        url:         `${API_URL}order/${orderId}/file/${type}`,
+        url:         `${API_URL}order/${order.id}/file/${type}`,
         headers:     {
           token: Auth.token,
         },
@@ -104,20 +82,21 @@ if ($form.length) {
         },
         fail:        (e, data) => {
           $progress.hide();
-l        }
+        }
       });
-
     });
   });
 
   function createFile(file) {
-    const $button = $('<button class="docs__delete"></button>');
     const $el = $(`<div class="docs__item">${file.originalFilename}</div>`);
-    $button.on('click', (e) => {
-      e.preventDefault();
-      API.deleteFile(file.filePath, Auth.token).then(() => $el.remove());
-    });
-    $el.append($button);
+    if (!isEdit) {
+      const $button = $('<button class="docs__delete"></button>');
+      $button.on('click', (e) => {
+        e.preventDefault();
+        API.deleteFile(file.filePath, Auth.token).then(() => $el.remove());
+      });
+      $el.append($button);
+    }
     return $el;
   }
-}
+};
